more cleanup and identity service implementation

+++ src/Application/Common/Models/JwtSettings.cs
@@ -5,6 +5,6 @@ public class JwtSettings
     public string Key { get; set; } = string.Empty;
     public string Issuer { get; set; } = string.Empty;
     public string Audience { get; set; } = string.Empty;
-    public double AccessTokenExpirationMinutes { get; set; } = 15;
-    public double RefreshTokenExpirationDays { get; set; } = 7;
+    public int AccessTokenExpirationMinutes { get; set; } = 15;
+    public int RefreshTokenExpirationDays { get; set; } = 7;
 }
